# internal/handlers/swagger_models.go
package handlers

type UserData struct {
	Username string `json:"username"`
	Password string `json:"password"`
}

type SuccessResponse struct {
	Message string `json:"message"`
}

type ErrorResponse struct {
	Error string `json:"error"`
}

type TaskData struct {
	Title       string `db:"title" json:"title"`
	Description string `db:"description" json:"description"`
	Status      string `db:"status" json:"status"`
}